`Added ARM template`
Added a new Azure Resource Manager (ARM) template in `profile.arm.json` for automating the deployment of resources in Azure. The template includes parameters for the resource group name, location, and the main resource name, and variables for the app service plan name and resource ID. The resources to be created include a resource group and a deployment, which includes a web app and a server farm. The web app is configured to use .NET Core and is associated with the server farm, which is configured with the Standard tier and size S1.