Added ProductsController and new pages
Added a new `ProductsController` class to the `CloudDevPOE.Controllers` namespace in the `ProductsController.cs` file. This class inherits from the `Controller` class and contains an `Index` method that returns a view. Also, several new pages have been added to the application, each with a corresponding `.cshtml` file. These pages include `HomeDecorPage`, `JewelryPage`, `MetalArtPage`, `PotteryPage`, `TextilesPage`, and `Woodwork`. Each page sets a title in the `ViewData` and contains a similar layout defined in CSS and HTML. The layout includes a grid container with six boxes, each with a different background color. The pages are styled with CSS, including a 4-column grid layout for the container and rounded edges for the boxes.