"Removed Main method and button, updated default resourceName"
Removed the `Main` method from the `HomeController.cs` file in the `CloudDevPOE.Controllers` namespace, which was responsible for rendering the Main view. Also, the `Main` button was removed from the `_Layout.cshtml` file, which was responsible for redirecting the user to the Main view. In the `profile.arm.json` file, the `resourceName` parameter's default value was changed from `KhumaloCraftArtisans` to `ST10269509`. This parameter is used to specify the name of the main resource to be created by the template.