`Updated user sign-up process and error handling`
Modified the `SignUp` method in `UserController.cs` to return a JSON object instead of a view. The method now checks if the user ID is returned after inserting a new user. If the user ID is returned, it is stored in the session and a JSON object with `success = true` is returned. If the user ID is not returned, a JSON object with `success = false` and an error message is returned. If the model state is not valid, a JSON object with `success = false` and an error message is returned.

Updated the `InsertUser` method in `Tbl_Users.cs` to return the user ID after inserting a new user. The SQL query has been updated to use the `OUTPUT INSERTED.user_id` clause to return the user ID. The `ExecuteNonQuery` method has been replaced with the `ExecuteScalar` method to retrieve the user ID. The method now returns `null` if the user ID is not returned.

The `showLoginModalAndSubmitForm` function in `loginSignup.js` has been updated to display a SweetAlert2 confirmation message when a user signs up successfully. The page is refreshed after the user closes the confirmation message. If the sign up is not successful, the error message is displayed in the error bar.