Updated user handling and form validation in views and scripts
In `ProductsController.cs`, the `userID` is now set in the `ViewData` dictionary using the session variable "UserId". This ID is then used in `AddProduct.cshtml` to validate form submission and prevent it if the `userID` is not set. The form submission event listener in `AddProduct.cshtml` has been updated to include this validation.

The image upload button in `AddProduct.cshtml` has been modified, but the change seems to be a duplicate and may not have any functional impact.

In `_Layout.cshtml`, a line of code responsible for displaying the main content of the page was removed, which might affect the layout and content display of the page.

Lastly, the validation for the signup form in `loginSignup.js` was modified to not hide the error message when the input length is 0, potentially improving user feedback during form validation.